requirements.txt
# SPDX-FileCopyrightText: 2021 Diego Elio Pettenò
#
# SPDX-License-Identifier: 0BSD

aioconsole
click
construct
pyserial
